requirements.txt
Flask==2.3.2
flask-socketio==5.3.2
eventlet==0.33.3
paho-mqtt==1.6.1
python-dotenv==1.0.0
